refactor(agentd): decompose the mvm-guest-agent binary into bin submodules
