+++ crates/mvm-agentd/src/bin/mvm-guest-agent/handlers.rs
@@ -1,0 +1,917 @@
+//! The always-compiled per-verb request handlers, plus the `RunEntrypoint`
+//! execution machinery (validated-entrypoint dispatch, warm-pool routing,
+//! and the per-call TMPDIR). Dev-only verbs (`Exec`, `RunCode`,
+//! `RunDetached`, the console) get a production stub here; their real
+//! implementation lives in `interactive` behind the `interactive` feature.
+
+use std::path::{Path, PathBuf};
+use std::sync::Arc;
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
+
+use mvm_agentd::entrypoint::{CallCaps, CallOutcome, PayloadCapStream, execute};
+use mvm_agentd::vsock::{
+    ComponentState, EntrypointEvent, FsChange, FsChangeKind, GuestResponse, RunEntrypointError,
+};
+use mvm_agentd::worker_pool::{DispatchError, DispatchOutcome, WorkerPool};
+use mvm_agentd::worker_protocol::WorkerOutcome;
+
+use crate::HandlerCtx;
+use crate::globals::{
+    RUN_ENTRYPOINT_LOCK, VALIDATED_ENTRYPOINT, WARM_POOL, reseed_on_post_restore,
+};
+use crate::health::build_integration_reports;
+use crate::port_forward::{run_port_forwarder, start_unix_socket_forwarder};
+use crate::probe::build_probe_reports;
+use crate::socket::write_response;
+
+/// Sync filesystems and drop page cache.
+fn do_sleep_prep() -> (bool, String) {
+    // Sync all filesystems.
+    let sync_ok = std::process::Command::new("sync")
+        .status()
+        .is_ok_and(|s| s.success());
+
+    // Drop page cache (requires root, best-effort).
+    let drop_ok = std::fs::write("/proc/sys/vm/drop_caches", "3").is_ok();
+
+    if sync_ok && drop_ok {
+        (true, "filesystems synced, page cache dropped".to_string())
+    } else if sync_ok {
+        (
+            true,
+            "filesystems synced, page cache drop failed (non-root?)".to_string(),
+        )
+    } else {
+        (false, "sync failed".to_string())
+    }
+}
+
+/// Generate a per-call TMPDIR path under /tmp. The mutex guarantees only
+/// one in-flight call per VM, so a name collision is exceedingly unlikely
+/// — but use pid + nanos anyway to survive any post-crash leftovers.
+fn make_call_tmpdir() -> std::io::Result<CallTmpdir> {
+    use std::os::unix::fs::DirBuilderExt;
+    let nanos = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .map(|d| d.as_nanos())
+        .unwrap_or(0);
+    let pid = std::process::id();
+    let path = PathBuf::from(format!("/tmp/mvm-call-{pid}-{nanos:x}"));
+    std::fs::DirBuilder::new().mode(0o700).create(&path)?;
+    Ok(CallTmpdir { path })
+}
+
+/// RAII wrapper that removes the TMPDIR on drop. The cleanup runs from the
+/// agent — robust to wrapper crashes, kills, and any panic on the agent's
+/// own side.
+struct CallTmpdir {
+    path: PathBuf,
+}
+
+impl CallTmpdir {
+    fn path(&self) -> &Path {
+        &self.path
+    }
+}
+
+impl Drop for CallTmpdir {
+    fn drop(&mut self) {
+        if let Err(e) = std::fs::remove_dir_all(&self.path) {
+            eprintln!(
+                "mvm-guest-agent: TMPDIR cleanup failed for {}: {e}",
+                self.path.display()
+            );
+        }
+    }
+}
+
+/// Wrap an `EntrypointEvent` in a `GuestResponse` for vsock framing.
+fn evt(e: EntrypointEvent) -> GuestResponse {
+    GuestResponse::EntrypointEvent(e)
+}
+
+/// Handle a `RunEntrypoint` request. Writes streaming events directly via
+/// `write_response` and returns the terminal event for the dispatcher to
+/// send through the existing `match` arm pattern.
+///
+/// `#[inline(never)]` is load-bearing for the symbol-contract gate.
+/// LTO inlines functions called from a single site, which would erase
+/// the `mvm_guest_agent::handle_run_entrypoint` symbol from `nm`
+/// output even though the handler is logically compiled in. The gate
+/// (`scripts/check-prod-agent-no-exec.sh`) requires the symbol to be
+/// present as positive evidence that the handler is wired up. Without
+/// `inline(never)` the gate fails on every prod build. Cost: one
+/// extra call boundary in a slow-path RPC handler — imperceptible.
+#[inline(never)]
+fn handle_run_entrypoint(
+    file: &mut std::fs::File,
+    stdin: Vec<u8>,
+    timeout_secs: u64,
+    env: Vec<(String, String)>,
+) -> GuestResponse {
+    // When a warm-process pool is active, route through it instead
+    // of the cold-respawn path. The host wire is identical;
+    // the pool's `dispatch` synthesizes the same `EntrypointEvent`
+    // stream (Stdout / Stderr / Exit | Error) we'd produce below.
+    if let Some(Some(pool)) = WARM_POOL.get() {
+        return dispatch_via_warm_pool(file, pool, stdin, timeout_secs, env);
+    }
+
+    let _guard = match RUN_ENTRYPOINT_LOCK.try_lock() {
+        Ok(g) => g,
+        Err(_) => {
+            return evt(EntrypointEvent::Error {
+                kind: RunEntrypointError::Busy,
+                message: "another RunEntrypoint call is in flight".into(),
+            });
+        }
+    };
+
+    let entrypoint = match VALIDATED_ENTRYPOINT.get() {
+        Some(Ok(e)) => e,
+        Some(Err(msg)) => {
+            return evt(EntrypointEvent::Error {
+                kind: RunEntrypointError::EntrypointInvalid,
+                message: msg.clone(),
+            });
+        }
+        None => {
+            return evt(EntrypointEvent::Error {
+                kind: RunEntrypointError::EntrypointInvalid,
+                message: "entrypoint validation never ran".into(),
+            });
+        }
+    };
+
+    let tmpdir = match make_call_tmpdir() {
+        Ok(t) => t,
+        Err(e) => {
+            return evt(EntrypointEvent::Error {
+                kind: RunEntrypointError::InternalError,
+                message: format!("create per-call TMPDIR: {e}"),
+            });
+        }
+    };
+
+    let outcome = execute(
+        entrypoint,
+        tmpdir.path(),
+        &stdin,
+        Duration::from_secs(timeout_secs),
+        CallCaps::v1(),
+        env,
+    );
+
+    // tmpdir drops at end of scope (or on early-return below) and runs
+    // its `Drop` cleanup.
+    match outcome {
+        CallOutcome::Exited {
+            code,
+            stdout,
+            stderr,
+            controls,
+        } => {
+            write_response(file, &evt(EntrypointEvent::Stdout { chunk: stdout }));
+            write_response(file, &evt(EntrypointEvent::Stderr { chunk: stderr }));
+            emit_controls(file, controls);
+            evt(EntrypointEvent::Exit { code })
+        }
+        CallOutcome::Timeout {
+            stdout,
+            stderr,
+            controls,
+        } => {
+            write_response(file, &evt(EntrypointEvent::Stdout { chunk: stdout }));
+            write_response(file, &evt(EntrypointEvent::Stderr { chunk: stderr }));
+            emit_controls(file, controls);
+            evt(EntrypointEvent::Error {
+                kind: RunEntrypointError::Timeout,
+                message: format!("wrapper exceeded {timeout_secs}s timeout"),
+            })
+        }
+        CallOutcome::PayloadCap {
+            stream,
+            stdout,
+            stderr,
+            controls,
+        } => {
+            write_response(file, &evt(EntrypointEvent::Stdout { chunk: stdout }));
+            write_response(file, &evt(EntrypointEvent::Stderr { chunk: stderr }));
+            emit_controls(file, controls);
+            let stream_name = match stream {
+                PayloadCapStream::Stdin => "stdin",
+                PayloadCapStream::Stdout => "stdout",
+                PayloadCapStream::Stderr => "stderr",
+            };
+            evt(EntrypointEvent::Error {
+                kind: RunEntrypointError::PayloadCap,
+                message: format!("{stream_name} exceeded its cap"),
+            })
+        }
+        CallOutcome::SpawnFailed { message } => evt(EntrypointEvent::Error {
+            kind: RunEntrypointError::InternalError,
+            message,
+        }),
+        CallOutcome::WrapperCrashed {
+            signal,
+            stdout,
+            stderr,
+            controls,
+        } => {
+            write_response(file, &evt(EntrypointEvent::Stdout { chunk: stdout }));
+            write_response(file, &evt(EntrypointEvent::Stderr { chunk: stderr }));
+            emit_controls(file, controls);
+            evt(EntrypointEvent::Error {
+                kind: RunEntrypointError::WrapperCrashed,
+                message: format!("wrapper exited via signal {signal}"),
+            })
+        }
+    }
+}
+
+/// Emit each fd-3 control record as one `EntrypointEvent::Control`
+/// frame on the response stream. The cold-path counterpart to the
+/// wrapper's own control emission. v1 emits all records after stdout
+/// and stderr; the host already accepts non-terminal events in any
+/// order before the terminal `Exit` / `Error`.
+fn emit_controls(file: &mut std::fs::File, records: Vec<mvm_agentd::entrypoint::ControlRecord>) {
+    for r in records {
+        write_response(
+            file,
+            &evt(EntrypointEvent::Control {
+                header_json: r.header_json,
+                payload: r.payload,
+            }),
+        );
+    }
+}
+
+/// Route a `RunEntrypoint` request through the warm-process worker
+/// pool. The pool's `dispatch` returns a single
+/// `DispatchOutcome` per call (one buffered frame), which we
+/// translate back into the existing host-facing `EntrypointEvent`
+/// stream — same wire shape as the cold path so `mvmctl invoke` is
+/// unaffected.
+///
+/// `#[inline(never)]` is load-bearing for the symbol-contract gate
+/// (mirrors the `handle_run_entrypoint` symbol invariant). The CI
+/// gate at `scripts/check-prod-agent-no-exec.sh`
+/// requires this symbol to be present on the same binary that
+/// ships, as positive evidence the warm-process substrate is
+/// wired in. Without `inline(never)` LTO would erase it from the
+/// `nm` output.
+#[inline(never)]
+fn dispatch_via_warm_pool(
+    file: &mut std::fs::File,
+    pool: &Arc<WorkerPool>,
+    stdin: Vec<u8>,
+    timeout_secs: u64,
+    env: Vec<(String, String)>,
+) -> GuestResponse {
+    match pool.dispatch(stdin, timeout_secs, env) {
+        Ok(DispatchOutcome {
+            stdout,
+            stderr,
+            controls,
+            outcome,
+        }) => {
+            write_response(file, &evt(EntrypointEvent::Stdout { chunk: stdout }));
+            write_response(file, &evt(EntrypointEvent::Stderr { chunk: stderr }));
+            emit_controls(file, controls);
+            match outcome {
+                WorkerOutcome::Exit { code } => evt(EntrypointEvent::Exit { code }),
+                WorkerOutcome::Error { kind, message } => evt(EntrypointEvent::Error {
+                    kind: map_worker_error_kind(&kind),
+                    message,
+                }),
+            }
+        }
+        Err(DispatchError::QueueFull) => evt(EntrypointEvent::Error {
+            kind: RunEntrypointError::Busy,
+            message: "warm-process pool queue is full".into(),
+        }),
+        Err(DispatchError::ShuttingDown) => evt(EntrypointEvent::Error {
+            kind: RunEntrypointError::InternalError,
+            message: "warm-process pool is shutting down".into(),
+        }),
+        Err(DispatchError::NoLiveWorkers) => evt(EntrypointEvent::Error {
+            kind: RunEntrypointError::InternalError,
+            message: "warm-process pool has no live workers".into(),
+        }),
+        // after_start readiness probe has not yet succeeded —
+        // workload is still warming up. Surface as Busy
+        // so the host's retry semantics apply (same as queue-full);
+        // the message distinguishes the cause for operators.
+        Err(DispatchError::NotReady) => evt(EntrypointEvent::Error {
+            kind: RunEntrypointError::Busy,
+            message: "warm-process pool warming up (after_start probe not yet ready)".into(),
+        }),
+    }
+}
+
+fn map_worker_error_kind(kind: &str) -> RunEntrypointError {
+    match kind {
+        "wrapper_crash" => RunEntrypointError::WrapperCrashed,
+        "timeout" => RunEntrypointError::Timeout,
+        _ => RunEntrypointError::InternalError,
+    }
+}
+
+/// Collect filesystem changes by walking the overlay upper directory.
+///
+/// When the rootfs is mounted read-only with an overlay (squashfs + tmpfs),
+/// all writes go to the upper dir (typically /overlay/upper). Walking it
+/// reveals every file created or modified since boot.
+///
+/// Falls back to an empty list if the overlay dir doesn't exist (non-overlay
+/// rootfs or unrestricted mode).
+fn collect_fs_diff() -> Vec<FsChange> {
+    // Common overlay upper dir paths
+    let upper_dirs = ["/overlay/upper", "/run/overlay/upper", "/tmp/overlay/upper"];
+    let upper = upper_dirs.iter().find(|p| std::path::Path::new(p).is_dir());
+
+    let Some(upper_dir) = upper else {
+        return Vec::new();
+    };
+
+    let mut changes = Vec::new();
+    walk_dir(std::path::Path::new(upper_dir), upper_dir, &mut changes);
+    changes.sort_by(|a, b| a.path.cmp(&b.path));
+    changes
+}
+
+fn walk_dir(dir: &std::path::Path, strip_prefix: &str, changes: &mut Vec<FsChange>) {
+    let Ok(entries) = std::fs::read_dir(dir) else {
+        return;
+    };
+    for entry in entries.flatten() {
+        let path = entry.path();
+        let rel = path
+            .to_str()
+            .unwrap_or("")
+            .strip_prefix(strip_prefix)
+            .unwrap_or("")
+            .to_string();
+
+        if rel.is_empty() {
+            continue;
+        }
+
+        if path.is_dir() {
+            walk_dir(&path, strip_prefix, changes);
+        } else {
+            let size = std::fs::metadata(&path).map(|m| m.len()).unwrap_or(0);
+            // In overlay upper dir, whiteout files (.wh.*) indicate deletion
+            let filename = path.file_name().and_then(|n| n.to_str()).unwrap_or("");
+            if let Some(deleted_name) = filename.strip_prefix(".wh.") {
+                let parent = path.parent().unwrap_or(&path);
+                let del_rel = parent
+                    .to_str()
+                    .unwrap_or("")
+                    .strip_prefix(strip_prefix)
+                    .unwrap_or("");
+                changes.push(FsChange {
+                    path: format!("{}/{}", del_rel, deleted_name),
+                    kind: FsChangeKind::Deleted,
+                    size: 0,
+                });
+            } else {
+                // File exists in upper = created or modified
+                changes.push(FsChange {
+                    path: rel,
+                    kind: FsChangeKind::Created, // can't distinguish create vs modify from overlay alone
+                    size,
+                });
+            }
+        }
+    }
+}
+
+pub(crate) fn handle_ping() -> GuestResponse {
+    GuestResponse::Pong
+}
+
+pub(crate) fn handle_worker_status(ctx: &mut HandlerCtx) -> GuestResponse {
+    let (status, last_busy_at) = match ctx.state.lock() {
+        Ok(s) => (s.status.clone(), s.last_busy_at.clone()),
+        Err(_) => ("unknown".to_string(), None),
+    };
+    GuestResponse::WorkerStatus {
+        status,
+        last_busy_at,
+    }
+}
+
+pub(crate) fn handle_sleep_prep(_drain_timeout_secs: u64) -> GuestResponse {
+    let (success, detail) = do_sleep_prep();
+    GuestResponse::SleepPrepAck {
+        success,
+        detail: Some(detail),
+    }
+}
+
+pub(crate) fn handle_wake(ctx: &mut HandlerCtx) -> GuestResponse {
+    // Reset monitoring state after wake from snapshot.
+    if let Ok(mut s) = ctx.state.lock() {
+        s.status = "idle".to_string();
+        s.last_busy_at = None;
+    }
+    GuestResponse::WakeAck { success: true }
+}
+
+pub(crate) fn handle_integration_status(ctx: &mut HandlerCtx) -> GuestResponse {
+    GuestResponse::IntegrationStatusReport {
+        integrations: build_integration_reports(ctx.integration_state, ctx.boot_state.boot_at),
+    }
+}
+
+pub(crate) fn handle_checkpoint_integrations(_integrations: Vec<String>) -> GuestResponse {
+    GuestResponse::CheckpointResult {
+        success: true,
+        failed: vec![],
+        detail: None,
+    }
+}
+
+pub(crate) fn handle_probe_status(ctx: &mut HandlerCtx) -> GuestResponse {
+    GuestResponse::ProbeStatusReport {
+        probes: build_probe_reports(ctx.probe_state),
+    }
+}
+
+pub(crate) fn handle_primed_status() -> GuestResponse {
+    GuestResponse::PrimedStatusReport {
+        primed: mvm_agentd::vsock::workload_is_primed_at(std::path::Path::new(
+            mvm_agentd::vsock::PRIMED_MARKER_PATH,
+        )),
+    }
+}
+
+pub(crate) fn handle_post_restore(
+    ctx: &mut HandlerCtx,
+    token: [u8; mvm_core::crypto::vmgenid::GENID_BYTES],
+    grant_envelope: Option<mvm_core::protocol::vm_backend::VerbGrantEnvelope>,
+) -> GuestResponse {
+    // First, rotate the VMGenID: feed the host-minted token to the
+    // process-resident reseeder. Its state is captured in the snapshot,
+    // so two clones of one snapshot both diverge from the captured
+    // value when the host delivers each a distinct fresh token. A
+    // zero token (no-rotation restore) is a no-op.
+    let reseeded = matches!(
+        reseed_on_post_restore(token),
+        mvm_agentd::genid::GenIdAction::Reseeded
+    );
+    // Re-pin the verb grant if the host sent a fresh envelope. This
+    // covers restore across a plan change (a fork mints a fresh
+    // host-signed grant with the child's new session_id/plan_nonce and
+    // may widen the verb set). The envelope is verified against the
+    // boot-pinned host-signer anchor — NOT the self-attested key inside
+    // the envelope, which any caller able to deliver a PostRestore could
+    // forge. With no boot anchor there is nothing to trust the envelope
+    // against, so the re-pin is refused (fail closed).
+    if let Some(env) = grant_envelope.as_ref() {
+        match ctx.boot_state.host_signer_key() {
+            Some(anchor) => {
+                let (_, current_grant) = ctx.boot_state.grant_state();
+                if let Some(g) = mvm_agentd::vsock::re_pin_verb_grant(
+                    env,
+                    current_grant.as_ref(),
+                    &anchor,
+                    chrono::Utc::now(),
+                ) {
+                    ctx.boot_state.set_verb_grant(g);
+                }
+            }
+            None => eprintln!(
+                "mvm-guest-agent: PostRestore re-pin refused — no boot-pinned host-signer anchor"
+            ),
+        }
+    }
+    // Then send SIGUSR1 to PID 1 to trigger drive remount + service restart.
+    let result = std::process::Command::new("kill")
+        .args(["-USR1", "1"])
+        .output();
+    match result {
+        Ok(out) if out.status.success() => GuestResponse::PostRestoreAck {
+            success: true,
+            detail: Some("post-restore signal sent to init".to_string()),
+            reseeded,
+        },
+        Ok(out) => GuestResponse::PostRestoreAck {
+            success: false,
+            detail: Some(format!(
+                "kill failed: {}",
+                String::from_utf8_lossy(&out.stderr)
+            )),
+            reseeded,
+        },
+        Err(e) => GuestResponse::PostRestoreAck {
+            success: false,
+            detail: Some(format!("failed to send signal: {}", e)),
+            reseeded,
+        },
+    }
+}
+
+#[cfg(not(feature = "interactive"))]
+pub(crate) fn handle_exec(
+    ctx: &mut HandlerCtx,
+    command: String,
+    stdin: Option<String>,
+    timeout_secs: Option<u64>,
+) -> GuestResponse {
+    let _ = (ctx, command, stdin, timeout_secs);
+    GuestResponse::Error {
+        message: "exec not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access".to_string(),
+    }
+}
+
+#[cfg(not(feature = "interactive"))]
+pub(crate) fn handle_exec_batch(
+    stages: Vec<mvm_agentd::vsock::StageFile>,
+    commands: Vec<Vec<String>>,
+    timeout_secs: Option<u64>,
+) -> GuestResponse {
+    let _ = (stages, commands, timeout_secs);
+    GuestResponse::Error {
+        message: "exec-batch not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+            .to_string(),
+    }
+}
+
+#[cfg(not(feature = "interactive"))]
+pub(crate) fn handle_run_code(
+    ctx: &mut HandlerCtx,
+    code: String,
+    timeout_secs: Option<u64>,
+) -> GuestResponse {
+    let _ = (ctx, code, timeout_secs);
+    GuestResponse::Error {
+        message: "run-code not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+            .to_string(),
+    }
+}
+
+/// `RunEntrypoint` arm: distinguish "validation hasn't completed yet"
+/// (Starting → NotReady, transient) from "validation failed"
+/// (Failed → EntrypointInvalid, terminal) before delegating to the
+/// real handler. Snapshot once so the decision is consistent even if
+/// a concurrent background-thread update flips state mid-handler.
+pub(crate) fn handle_run_entrypoint_request(
+    ctx: &mut HandlerCtx,
+    stdin: Vec<u8>,
+    timeout_secs: u64,
+    env: Vec<(String, String)>,
+) -> GuestResponse {
+    if matches!(
+        ctx.boot_state.snapshot().entrypoint,
+        ComponentState::Starting
+    ) {
+        GuestResponse::EntrypointEvent(EntrypointEvent::Error {
+            kind: RunEntrypointError::NotReady,
+            message: "entrypoint validation in progress; poll ReadinessStatus and retry"
+                .to_string(),
+        })
+    } else {
+        handle_run_entrypoint(ctx.file, stdin, timeout_secs, env)
+    }
+}
+
+#[cfg(not(feature = "interactive"))]
+pub(crate) fn handle_run_detached(argv: Vec<String>, env: Vec<(String, String)>) -> GuestResponse {
+    let _ = (argv, env);
+    GuestResponse::Error {
+        message: "run-detached not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+            .to_string(),
+    }
+}
+
+pub(crate) fn handle_fs_diff() -> GuestResponse {
+    // Walk the overlay upper dir to find changes since boot.
+    // The overlay upper dir is typically at /overlay/upper when
+    // the rootfs is mounted read-only with an overlay.
+    let changes = collect_fs_diff();
+    GuestResponse::FsDiffResult { changes }
+}
+
+pub(crate) fn handle_start_port_forward(guest_port: u16) -> GuestResponse {
+    let vsock_port = mvm_agentd::vsock::PORT_FORWARD_BASE + guest_port as u32;
+    eprintln!("port-fwd: starting vsock:{vsock_port} → tcp://localhost:{guest_port}");
+    std::thread::spawn(move || {
+        run_port_forwarder(vsock_port, guest_port);
+    });
+    GuestResponse::PortForwardStarted {
+        guest_port,
+        vsock_port,
+    }
+}
+
+pub(crate) fn handle_start_unix_socket_forward(
+    guest_path: String,
+    host_vsock_port: u32,
+    socket_mode: u32,
+) -> GuestResponse {
+    match start_unix_socket_forwarder(&guest_path, host_vsock_port, socket_mode) {
+        Ok(()) => GuestResponse::UnixSocketForwardStarted {
+            guest_path,
+            host_vsock_port,
+        },
+        Err(err) => GuestResponse::Error {
+            message: format!("unix socket forward failed: {err}"),
+        },
+    }
+}
+
+#[cfg(not(feature = "interactive"))]
+pub(crate) fn handle_console_open(
+    cols: u16,
+    rows: u16,
+    env: Vec<(String, String)>,
+    argv: Vec<String>,
+) -> GuestResponse {
+    let _ = (cols, rows, env, argv);
+    GuestResponse::Error {
+        message: "console not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+            .to_string(),
+    }
+}
+
+#[cfg(not(feature = "interactive"))]
+pub(crate) fn handle_console_close(session_id: u32) -> GuestResponse {
+    let _ = session_id;
+    GuestResponse::Error {
+        message: "console not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+            .to_string(),
+    }
+}
+
+#[cfg(not(feature = "interactive"))]
+pub(crate) fn handle_console_resize(session_id: u32, cols: u16, rows: u16) -> GuestResponse {
+    let _ = (session_id, cols, rows);
+    GuestResponse::Error {
+        message: "console not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+            .to_string(),
+    }
+}
+
+pub(crate) fn handle_entrypoint_status() -> GuestResponse {
+    match VALIDATED_ENTRYPOINT.get() {
+        Some(Ok(v)) => GuestResponse::EntrypointStatusReport {
+            ok: true,
+            path: Some(v.resolved.display().to_string()),
+            detail: None,
+        },
+        Some(Err(msg)) => GuestResponse::EntrypointStatusReport {
+            ok: false,
+            path: None,
+            detail: Some(msg.clone()),
+        },
+        // With background init, `None` means validation is still
+        // running, not "never ran".
+        None => GuestResponse::EntrypointStatusReport {
+            ok: false,
+            path: None,
+            detail: Some("entrypoint validation in progress".to_string()),
+        },
+    }
+}
+
+pub(crate) fn handle_readiness_status(ctx: &mut HandlerCtx) -> GuestResponse {
+    GuestResponse::ReadinessStatusReport(ctx.boot_state.snapshot())
+}
+
+pub(crate) fn handle_fs_read(path: String, offset: Option<u64>, length: u64) -> GuestResponse {
+    GuestResponse::FsResult(mvm_agentd::fs_rpc::handle_with_defaults(
+        mvm_agentd::fs_rpc::FsRequest::Read {
+            path: &path,
+            offset,
+            length,
+        },
+    ))
+}
+
+pub(crate) fn handle_fs_write(
+    path: String,
+    content: Vec<u8>,
+    mode: u32,
+    create_parents: bool,
+) -> GuestResponse {
+    GuestResponse::FsResult(mvm_agentd::fs_rpc::handle_with_defaults(
+        mvm_agentd::fs_rpc::FsRequest::Write {
+            path: &path,
+            content: &content,
+            mode,
+            create_parents,
+        },
+    ))
+}
+
+pub(crate) fn handle_fs_list(path: String) -> GuestResponse {
+    GuestResponse::FsResult(mvm_agentd::fs_rpc::handle_with_defaults(
+        mvm_agentd::fs_rpc::FsRequest::List { path: &path },
+    ))
+}
+
+pub(crate) fn handle_fs_stat(path: String, follow_symlinks: bool) -> GuestResponse {
+    GuestResponse::FsResult(mvm_agentd::fs_rpc::handle_with_defaults(
+        mvm_agentd::fs_rpc::FsRequest::Stat {
+            path: &path,
+            follow_symlinks,
+        },
+    ))
+}
+
+pub(crate) fn handle_fs_mkdir(path: String, mode: u32, parents: bool) -> GuestResponse {
+    GuestResponse::FsResult(mvm_agentd::fs_rpc::handle_with_defaults(
+        mvm_agentd::fs_rpc::FsRequest::Mkdir {
+            path: &path,
+            mode,
+            parents,
+        },
+    ))
+}
+
+pub(crate) fn handle_fs_remove(path: String, recursive: bool) -> GuestResponse {
+    GuestResponse::FsResult(mvm_agentd::fs_rpc::handle_with_defaults(
+        mvm_agentd::fs_rpc::FsRequest::Remove {
+            path: &path,
+            recursive,
+        },
+    ))
+}
+
+pub(crate) fn handle_fs_move(from: String, to: String) -> GuestResponse {
+    GuestResponse::FsResult(mvm_agentd::fs_rpc::handle_with_defaults(
+        mvm_agentd::fs_rpc::FsRequest::Move {
+            from: &from,
+            to: &to,
+        },
+    ))
+}
+
+pub(crate) fn handle_proc_start(
+    argv: Vec<String>,
+    env: std::collections::BTreeMap<String, String>,
+    cwd: Option<String>,
+    stdin: Vec<u8>,
+) -> GuestResponse {
+    #[cfg(feature = "interactive")]
+    {
+        let caps = mvm_agentd::process_rpc::Caps::production();
+        GuestResponse::ProcResult(mvm_agentd::process_rpc::handle_proc_start(
+            crate::interactive::proc_registry(),
+            &caps,
+            &argv,
+            &env,
+            cwd.as_deref(),
+            &stdin,
+        ))
+    }
+    #[cfg(not(feature = "interactive"))]
+    {
+        let _ = (argv, env, cwd, stdin);
+        GuestResponse::ProcResult(mvm_agentd::vsock::ProcResult::Error {
+            kind: mvm_agentd::vsock::ProcErrorKind::UnsupportedInProduction,
+            message:
+                "process control not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+                    .to_string(),
+        })
+    }
+}
+
+pub(crate) fn handle_proc_list() -> GuestResponse {
+    #[cfg(feature = "interactive")]
+    {
+        GuestResponse::ProcResult(mvm_agentd::process_rpc::handle_proc_list(
+            crate::interactive::proc_registry(),
+        ))
+    }
+    #[cfg(not(feature = "interactive"))]
+    {
+        GuestResponse::ProcResult(mvm_agentd::vsock::ProcResult::Error {
+            kind: mvm_agentd::vsock::ProcErrorKind::UnsupportedInProduction,
+            message:
+                "process control not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+                    .to_string(),
+        })
+    }
+}
+
+pub(crate) fn handle_proc_signal(pid_token: String, signum: i32) -> GuestResponse {
+    #[cfg(feature = "interactive")]
+    {
+        GuestResponse::ProcResult(mvm_agentd::process_rpc::handle_proc_signal(
+            crate::interactive::proc_registry(),
+            &pid_token,
+            signum,
+        ))
+    }
+    #[cfg(not(feature = "interactive"))]
+    {
+        let _ = (pid_token, signum);
+        GuestResponse::ProcResult(mvm_agentd::vsock::ProcResult::Error {
+            kind: mvm_agentd::vsock::ProcErrorKind::UnsupportedInProduction,
+            message:
+                "process control not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+                    .to_string(),
+        })
+    }
+}
+
+pub(crate) fn handle_proc_send_input(pid_token: String, bytes: Vec<u8>) -> GuestResponse {
+    #[cfg(feature = "interactive")]
+    {
+        let caps = mvm_agentd::process_rpc::Caps::production();
+        GuestResponse::ProcResult(mvm_agentd::process_rpc::handle_proc_send_input(
+            crate::interactive::proc_registry(),
+            &caps,
+            &pid_token,
+            &bytes,
+        ))
+    }
+    #[cfg(not(feature = "interactive"))]
+    {
+        let _ = (pid_token, bytes);
+        GuestResponse::ProcResult(mvm_agentd::vsock::ProcResult::Error {
+            kind: mvm_agentd::vsock::ProcErrorKind::UnsupportedInProduction,
+            message:
+                "process control not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+                    .to_string(),
+        })
+    }
+}
+
+pub(crate) fn handle_proc_kill(pid_token: String) -> GuestResponse {
+    #[cfg(feature = "interactive")]
+    {
+        GuestResponse::ProcResult(mvm_agentd::process_rpc::handle_proc_kill(
+            crate::interactive::proc_registry(),
+            &pid_token,
+        ))
+    }
+    #[cfg(not(feature = "interactive"))]
+    {
+        let _ = pid_token;
+        GuestResponse::ProcResult(mvm_agentd::vsock::ProcResult::Error {
+            kind: mvm_agentd::vsock::ProcErrorKind::UnsupportedInProduction,
+            message:
+                "process control not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+                    .to_string(),
+        })
+    }
+}
+
+pub(crate) fn handle_proc_wait(
+    ctx: &mut HandlerCtx,
+    pid_token: String,
+    timeout_secs: Option<u64>,
+) -> GuestResponse {
+    #[cfg(feature = "interactive")]
+    {
+        let terminal =
+            crate::interactive::handle_proc_wait_streaming(ctx.file, &pid_token, timeout_secs);
+        GuestResponse::ProcWaitEvent(terminal)
+    }
+    #[cfg(not(feature = "interactive"))]
+    {
+        let _ = (ctx, pid_token, timeout_secs);
+        GuestResponse::ProcWaitEvent(mvm_agentd::vsock::ProcWaitEvent::Error {
+            kind: mvm_agentd::vsock::ProcErrorKind::UnsupportedInProduction,
+            message:
+                "process control not available on a production workload: this is a dev-only control verb, absent from sealed agents — run a dev-tier workload for interactive/exec access"
+                    .to_string(),
+        })
+    }
+}
+
+pub(crate) fn handle_mount_volume(
+    volume_name: String,
+    guest_path: String,
+    read_only: bool,
+) -> GuestResponse {
+    GuestResponse::VolumeMountResult(mvm_agentd::volume::handle_mount(
+        &volume_name,
+        &guest_path,
+        read_only,
+    ))
+}
+
+pub(crate) fn handle_unmount_volume(guest_path: String, force: bool) -> GuestResponse {
+    GuestResponse::VolumeMountResult(mvm_agentd::volume::handle_unmount(&guest_path, force))
+}
+
+pub(crate) fn handle_update_idle_timeout(secs: u64) -> GuestResponse {
+    match WARM_POOL.get() {
+        Some(Some(pool)) => {
+            let previous = pool.set_idle_timeout(secs);
+            GuestResponse::UpdateIdleTimeoutAck {
+                previous_secs: previous,
+                applied_secs: secs,
+            }
+        }
+        _ => GuestResponse::UpdateIdleTimeoutAck {
+            previous_secs: 0,
+            applied_secs: 0,
+        },
+    }
+}
